Fix type of variable read from command line

+++ synchrono/intersection.cpp
@@ -67,7 +67,7 @@ double step_size = 3e-3;
 double render_step_size = 1.0 / 50;  // FPS = 50
 
 // How often SynChrono state messages are interchanged
-float heartbeat = 1e-2f;  // 100[Hz]
+double heartbeat = 1e-2;  // 100[Hz]
 
 void AddCommandLineOptions(ChCLI& cli);
 void LogCopyright(bool show);
@@ -103,7 +103,7 @@ int main(int argc, char* argv[]) {
     // Normal simulation options
     step_size = cli.GetAsType<double>("step_size");
     end_time = cli.GetAsType<double>("end_time");
-    heartbeat = cli.GetAsType<float>("heartbeat");
+    heartbeat = cli.GetAsType<double>("heartbeat");
 
     const bool use_sensor = cli.HasValueInVector<int>("sens", node_id);
     const bool sensor_vis = cli.GetAsType<bool>("sens_vis");
